add div grad
