+++ tfjs-core/src/ops/ops.ts
@@ -17,6 +17,7 @@
 
 // Modularized ops.
 export {broadcastTo} from './broadcast_to';
+export {div} from './div';
 export {square} from './square';
 export {squaredDifference} from './squared_difference';
