+++ tfjs-core/src/register_all_gradients.ts
@@ -15,6 +15,7 @@
  * =============================================================================
  */
 import {broadcastToGradConfig} from './gradients/BroadcastTo_grad';
+import {divGradConfig} from './gradients/Div_grad';
 import {squareGradConfig} from './gradients/Square_grad';
 import {squaredDifferenceGradConfig} from './gradients/SquaredDifference_grad';
 import {GradConfig} from './kernel_registry';
@@ -24,6 +25,7 @@ import {registerGradient} from './kernel_registry';
 const gradConfigs: GradConfig[] = [
   squareGradConfig,
   squaredDifferenceGradConfig,
+  divGradConfig,
   broadcastToGradConfig,
 ];
